Add automated deployment of Grafana dashboards ConfigMap
Introduces a ConfigMap for Grafana dashboards with both static (k8s) and templated (Ansible) approaches. Updates Ansible and setup scripts to deploy and clean up the ConfigMap, and configures Grafana Helm values to enable dashboard sidecar loading. This enables automated provisioning and management of application and security dashboards in the monitoring namespace.